feat: metrics for outboxed services (#339)
Add collection of queue metrics - unless the feature is disabled - by
registering a `done` handler, which will trigger metrics collection.
Metrics collection for a queued service will start once the service in
question is found to be the `target` of a queued message for the first
time.

Correct collection of the `cold_entries` metric depends on the
`maxAttempts` queue configuration being available from the queued
service, which is not yet the case for cds^8. For cds^8 `maxAttempts` is
currently being defaulted (to 20).

Counts of `incoming_messages` and `outgoing_messages` are currently
being collected per tenant per app instance and AFAIK should be
aggregated per tenant by the collector.

---------

Co-authored-by: sjvans <30337871+sjvans@users.noreply.github.com>
Co-authored-by: D050513 <sebastian.van.syckel@sap.com>